Save opam's build cache on post
This significantly speeds up the build time of dependencies
